+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 - Storage Templates
 - Storage Volume Attachments
 - Storage Volume
+- Server Profiles
 - Uplink set
 
 # 5.1.1